+++ client/config/webpack/webpack.config.dev.js
@@ -1,8 +1,9 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-
 const path = require("path")
+const dotEnv = require("dotenv-webpack")
+
 const shared = require("./webpack.config.shared")
-const rootFolder = path.resolve(__dirname, "..")
+const { ROOT_FOLDER } = require("./helpers")
 
 /**
  * See [Webpack Configuration docs](https://webpack.js.org/configuration/) for more information.
@@ -17,10 +18,10 @@ const webpackConfig = {
     devServer: {
         static: [
             {
-                directory: path.join(rootFolder, "dist"),
+                directory: path.join(ROOT_FOLDER, "dist"),
             },
             {
-                directory: path.join(rootFolder, "../etc/assets"),
+                directory: path.join(ROOT_FOLDER, "../etc/assets"),
                 publicPath: "/public",
             },
         ],
@@ -29,6 +30,12 @@ const webpackConfig = {
         allowedHosts: "all",
         hot: true,
     },
+    plugins: [
+        ...shared.plugins,
+        new dotEnv({
+            path: path.resolve(ROOT_FOLDER, ".env.dev"),
+        }),
+    ],
 }
 
 module.exports = webpackConfig