+++ client/config/webpack/webpack.config.prod.js
@@ -1,6 +1,9 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
+const path = require("path")
+const dotEnv = require("dotenv-webpack")
 
 const shared = require("./webpack.config.shared")
+const { ROOT_FOLDER } = require("./helpers")
 
 /**
  * See [Webpack Configuration docs](https://webpack.js.org/configuration/) for more information.
@@ -10,6 +13,12 @@ const shared = require("./webpack.config.shared")
 const webpackConfig = {
     ...shared,
     mode: "production",
+    plugins: [
+        ...shared.plugins,
+        new dotEnv({
+            path: path.resolve(ROOT_FOLDER, ".env.prod"),
+        }),
+    ],
 }
 
 module.exports = webpackConfig